refactored tests.py
